fix(index): Prevent Phantom Vectors and improve HNSW coverage
- Fix Phantom Vector race condition in `HnswIndex` by enforcing strict lock ordering (Inner -> DashMap).
- Fix deadlock potential in `add` (Occupied path) using optimistic concurrency.
- Fix atomicity gap in `add` (Vacant path) by holding lock during ID insertion.
- Reorder `save_internal` to prioritize Index snapshot over Mappings.
- Move coverage tests to `src/index/vector/hnsw.rs` to ensure proper instrumentation.
- Add `test_havoc_save_consistency.rs` for regression testing.
- Add unit tests for retry loops and error paths in `HnswIndex`.
